Removal of DTO package and refactor related to that
