--- server/src/main/kotlin/com/discover/server/dto/Feed.kt
@@ -1,6 +1,0 @@
-package com.discover.server.dto
-
-data class Feed(val title: String,
-                val summary: String,
-                val url: String,
-                val timeCreated: String)
