--- server/src/main/kotlin/com/discover/server/dto/Source.kt
@@ -1,5 +1,0 @@
-package com.discover.server.dto
-
-import javax.validation.constraints.NotEmpty
-
-data class SourceRequest(@field:NotEmpty val url: String)
