+++ server/src/main/kotlin/com/discover/server/facade/SourceFacade.kt
@@ -1,9 +1,9 @@
 package com.discover.server.facade
 
-import com.discover.server.dto.SearchCriteria
-import com.discover.server.dto.SourceRequest
-import com.discover.server.model.Source
-import com.discover.server.model.User
+import com.discover.server.domain.SearchCriteria
+import com.discover.server.domain.Source
+import com.discover.server.domain.SourceDTO
+import com.discover.server.domain.User
 import com.discover.server.service.SearchService
 import com.discover.server.service.SourceService
 import ma.glasnost.orika.MapperFacade
@@ -14,17 +14,17 @@ class SourceFacade(private val mapper: MapperFacade,
                    private val sourceService: SourceService,
                    private val searchService: SearchService<Source>) {
 
-    fun addSource(sourceRequest: SourceRequest, user: User): SourceRequest {
+    fun addSource(sourceRequest: SourceDTO, user: User): SourceDTO {
         val source = mapper.map(sourceRequest, Source::class.java)
         val addedSource = sourceService.addSource(source, user)
-        return mapper.map(addedSource, SourceRequest::class.java)
+        return mapper.map(addedSource, SourceDTO::class.java)
     }
 
     fun getSources() = sourceService.getSources()
 
     fun getSource(id: String) = sourceService.getSource(id)
 
-    fun updateSource(id: String, sourceRequest: SourceRequest) {
+    fun updateSource(id: String, sourceRequest: SourceDTO) {
         val source = mapper.map(sourceRequest, Source::class.java)
         sourceService.updateSource(id, source)
     }
@@ -33,9 +33,9 @@ class SourceFacade(private val mapper: MapperFacade,
         sourceService.deleteSourceById(id)
     }
 
-    fun findAll(searchCriteria: SearchCriteria): List<SourceRequest> {
+    fun findAll(searchCriteria: SearchCriteria): List<SourceDTO> {
         val specifications = searchCriteria.criteria.map { searchService.getSearchPredicate(it) }.toSet()
         val matchingSources = sourceService.findAll(specifications)
-        return mapper.mapAsList(matchingSources, SourceRequest::class.java)
+        return mapper.mapAsList(matchingSources, SourceDTO::class.java)
     }
 }
