+++ server/src/main/kotlin/com/discover/server/repository/RoleRepository.kt
@@ -1,6 +1,6 @@
 package com.discover.server.repository
 
-import com.discover.server.model.Role
+import com.discover.server.domain.Role
 import org.springframework.data.repository.CrudRepository
 
 interface RoleRepository: CrudRepository<Role, Long>
