+++ server/src/main/kotlin/com/discover/server/repository/UserRepository.kt
@@ -1,6 +1,6 @@
 package com.discover.server.repository
 
-import com.discover.server.model.User
+import com.discover.server.domain.User
 import org.springframework.data.jpa.repository.JpaRepository
 
 interface UserRepository: JpaRepository<User, Long> {
