+++ server/src/main/kotlin/com/discover/server/service/SourceService.kt
@@ -1,7 +1,7 @@
 package com.discover.server.service
 
-import com.discover.server.model.Source
-import com.discover.server.model.User
+import com.discover.server.domain.Source
+import com.discover.server.domain.User
 import com.discover.server.repository.SourceRepository
 import org.springframework.data.jpa.domain.Specification
 import org.springframework.stereotype.Service
