+++ server/src/test/kotlin/com/discover/server/service/SearchServiceTest.kt
@@ -1,8 +1,8 @@
 package com.discover.server.service
 
-import com.discover.server.dto.Criteria
-import com.discover.server.dto.Operator
-import com.discover.server.model.Source
+import com.discover.server.domain.Criteria
+import com.discover.server.domain.Operator
+import com.discover.server.domain.Source
 import io.mockk.every
 import io.mockk.mockk
 import io.mockk.verify
